redistributed assets
