+++ components/Terminal.tsx
@@ -1,9 +1,10 @@
-import { Col, Container, Row } from "react-bootstrap";
+import { Col, Container, Row, Image } from "react-bootstrap";
 
 import Heading from "./Heading";
 import Link from "./Link";
 import React from "react";
 import styled from "styled-components";
+import media from "../theme/media";
 
 const StyledFrame = styled(Col)`
   background-image: linear-gradient(
@@ -132,6 +133,15 @@ type TerminalHistoryProps = {
   id: string;
 };
 
+const Trail = styled(Image)`
+  position: absolute;
+  top: 10%;
+  z-index: -1;
+  display: none;
+  ${media.lg`width: 100%!important;`};
+  ${media.lg`display: block;`};
+`;
+
 const Terminal = ({ heading, commands, link }: TerminalProps) => {
   const handleCommandSubmit = (command: string): void => {
     let value = "Unknown command.";
@@ -183,39 +193,48 @@ const Terminal = ({ heading, commands, link }: TerminalProps) => {
   });
 
   return (
-    <Container>
-      <Row className="justify-content-center">
-        <StyledFrame className="mb-4">
-          <StyledTerminal>
-            <div className="w-100 d-flex justify-content-center align-items-center mb-3">
-              <StyledHeading>{heading}</StyledHeading>
-            </div>
-            <StyledHistoryWrapper ref={R}>
-              {terminalHistory.map((item, i) =>
-                item.isPrompt ? (
-                  <Prompt
-                    key={i}
-                    active={item.active}
-                    val={item.val}
-                    submitCommand={handleCommandSubmit}
-                    id={item.id}
-                  />
-                ) : (
-                  <StyledCommandResult key={i} id={item.id}>
-                    {item.val}
-                  </StyledCommandResult>
-                )
-              )}
-            </StyledHistoryWrapper>
-          </StyledTerminal>
-        </StyledFrame>
-      </Row>
-      <Row className="justify-content-center">
-        <Col style={{ maxWidth: 850 }}>
-          <Link href={link.href}>{link.label}</Link>
-        </Col>
-      </Row>
-    </Container>
+    <>
+      <div>
+        <Trail
+          src="img/section-graphic/trail.png"
+          alt={heading + " Graphic"}
+          fluid
+        />
+        <Container>
+          <Row className="justify-content-center">
+            <StyledFrame className="mb-4">
+              <StyledTerminal>
+                <div className="w-100 d-flex justify-content-center align-items-center mb-3">
+                  <StyledHeading>{heading}</StyledHeading>
+                </div>
+                <StyledHistoryWrapper ref={R}>
+                  {terminalHistory.map((item, i) =>
+                    item.isPrompt ? (
+                      <Prompt
+                        key={i}
+                        active={item.active}
+                        val={item.val}
+                        submitCommand={handleCommandSubmit}
+                        id={item.id}
+                      />
+                    ) : (
+                      <StyledCommandResult key={i} id={item.id}>
+                        {item.val}
+                      </StyledCommandResult>
+                    )
+                  )}
+                </StyledHistoryWrapper>
+              </StyledTerminal>
+            </StyledFrame>
+          </Row>
+          <Row className="justify-content-center">
+            <Col style={{ maxWidth: 850 }}>
+              <Link href={link.href}>{link.label}</Link>
+            </Col>
+          </Row>
+        </Container>
+      </div>
+    </>
   );
 };
